[psantos] Rebasing with master and adding faker packege within the specs
