+++ test/functional/UserController/displayUser.spec.js
@@ -2,35 +2,33 @@
 
 const { test, trait, before, after } = use('Test/Suite')('UserController');
 const User = use('App/Models/User');
+const faker = require("faker");
 
 trait('Test/ApiClient');
 trait('DatabaseTransactions');
 
-before(async () => {
+test('return the user based on its ID', async ({ client, assert }) => {
+  const uuid = faker.random.uuid();
+  const userName = faker.internet.userName();
+  const email = faker.internet.email();
+
   await User.create({
-    id: 123,
-    username: 'displayUser',
-    email: 'displayuser@user.com',
+    id: uuid,
+    username: userName,
+    email: email,
     password: '123456'
   });
-});
 
-after(async () => {
-  const user = await User.findOrFail(123);
-  await user.delete();
-});
-
-test('return the user based on its ID', async ({ client, assert }) => {
   assert.plan(2);
 
-  const response = await client.get(`/users/${123}`)
+  const response = await client.get(`/user/${uuid}`)
     .header('accept', 'application/json')
     .end();
   
   response.assertStatus(200);
   response.assertJSONSubset({
-    id: 123,
-    username: 'displayUser',
-    email: 'displayuser@user.com'
+    id: uuid,
+    username: userName,
+    email: email,
   });
 });
