+++ test/functional/UserController/updateUser.spec.js
@@ -2,28 +2,26 @@
 
 const { test, trait, before, after } = use('Test/Suite')('UserController');
 const User = use('App/Models/User');
+const faker = require("faker");
 
 trait('Test/ApiClient');
 trait('DatabaseTransactions');
 
-before(async () => {
+test('updates a new user', async ({ client, assert }) => {
+  const uuid = faker.random.uuid();
+  const userName = faker.internet.userName();
+  const email = faker.internet.email();
+
   await User.create({
-    id: 123,
-    username: 'updateUser',
-    email: 'updateuser@user.com',
+    id: uuid,
+    username: userName,
+    email: email,
     password: '123456'
   });
-});
 
-after(async () => {
-  const user = await User.findOrFail(123);
-  await user.delete();
-});
-
-test('updates a new user', async ({ client, assert }) => {
   assert.plan(2);
 
-  const response = await client.put(`/users/${123}`)
+  const response = await client.put(`user/${uuid}`)
     .send({
       username: 'userHasBeenUpdated',
       email: 'userhasbeenupdated@user.com',
@@ -32,8 +30,7 @@ test('updates a new user', async ({ client, assert }) => {
 
   response.assertStatus(200);
   response.assertJSONSubset({
-    id: 123,
     username: 'userHasBeenUpdated',
-    email: 'userhasbeenupdated@user.com'
+    email: 'userhasbeenupdated@user.com',
   });
 });
